move tests to protocol_test.rs
